Fix integer type casting in menu item iteration for ExitScreen and MainMenuScreen

+++ PSP Project/Src/ExitScreen.cpp
@@ -30,7 +30,7 @@ void ExitScreen::draw()
         {Resource::STR_NO, 160}
     };
 
-    for (int i = 0; i < sizeof(menuItems) / sizeof(menuItems[0]); ++i)
+    for (int i = 0; i < (int)(sizeof(menuItems) / sizeof(menuItems[0])); ++i)
     {
         if (gMenu == i)
         {
