+++ PSP Project/Src/MainMenuScreen.cpp
@@ -39,7 +39,7 @@ void MainMenuScreen::draw()
         {Resource::STR_EXIT_CAPTION, 200}
     };
 
-    for (int i = 0; i < sizeof(menuItems)/sizeof(menuItems[0]); ++i)
+    for (int i = 0; i < (int)(sizeof(menuItems)/sizeof(menuItems[0])); ++i)
     {
         if (gMenu == i)
         {
